Improvements for sharing files with other apps and receiving files from other apps. Better handling of Dialog question listeners to allow serializable support where possible.

+++ app/src/main/java/delit/piwigoclient/ui/events/trackable/TrackableRequestEvent.java
@@ -17,7 +17,7 @@ public class TrackableRequestEvent implements Serializable {
     }
 
     public synchronized static int getNextEventId() {
-        int id = actionIdGenerator.incrementAndGet();
+        short id = (short) actionIdGenerator.incrementAndGet();
         if (id < 0) {
             id = 0;
             actionIdGenerator.set(0);
